# docs/openapi/hansapp-openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/oauth/token": {
      "post": {
        "description": "grant_type=authorization_code 는 인가코드(code)를 토큰으로 교환한다. 코드에 기록된 클라이언트의 등록 오리진과 요청 Origin 이 일치해야 한다.\n\ngrant_type=refresh_token 은 refresh token(바디 또는 httpOnly 쿠키)을 갱신한다. **refresh token 은 1회용이다** — 갱신에 성공하면 새 refresh token 이 나오고 **직전 것은 그 즉시 무효**가 된다(rotate). 응답으로 받은 새 값으로 반드시 교체해 두고, 이전 값을 다시 보내면 거부된다. 같은 토큰으로 두 번 갱신하거나, 갱신 응답을 저장하지 못한 채 재시도하면 세션이 끊어진다.",
        "operationId": "OAuthController_token",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponseDto"
                }
              }
            }
          }
        },
        "summary": "토큰 발급/갱신",
        "tags": [
          "auth"
        ]
      }
    },
    "/oauth/authorize": {
      "post": {
        "description": "OAuth 2.0 Authorization Code Grant 의 인가 코드(authorization code)를 발급한다. 로그인된 사용자에 대해, 클라이언트가 등록한 `redirect_uri` 로 전달할 1회용 코드다.\n\n발급된 코드는 `POST /oauth/token` 에서 access token 으로 교환한다.",
        "operationId": "OAuthController_authorize",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorizeRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorizeResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "인가 코드 발급",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/users/me": {
      "get": {
        "description": "현재 로그인 사용자 정보를 반환한다.",
        "operationId": "UserController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "내 정보",
        "tags": [
          "users"
        ]
      }
    },
...
  },
  "info": {
    "title": "Hans API",
    "description": "Hans API backend 문서",
    "version": "0.15.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://127.0.0.1:3000",
      "description": "Local"
    },
    {
      "url": "https://develop-api.plzhans.com",
      "description": "Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT or service key(sk_...)",
        "type": "http",
        "description": "사용자 access token(JWT) 또는 서비스 키(sk_{appId}_{keyId}_{rand})."
      },
      "client-id": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Client-Id",
        "description": "공개 클라이언트 식별자(clientId). WEB 은 Origin 도 검사됨."
      }
    },
    "schemas": {
...
      "TokenRequestDto": {
        "type": "object",
        "properties": {
          "grant_type": {
            "type": "string",
            "description": "grant 종류",
            "enum": [
              "authorization_code",
              "refresh_token"
            ],
            "example": "authorization_code"
          },
          "code": {
            "type": "string",
            "description": "authorization_code grant 의 인가코드(ac_...)"
          },
          "refresh_token": {
            "type": "string",
            "description": "refresh_token grant 의 refresh token(rt_...). 생략 시 httpOnly 쿠키에서 읽는다."
          },
          "code_verifier": {
            "type": "string",
            "description": "PKCE code_verifier(RFC 7636). authorization_code grant 에 **필수**다. 인가 요청 때 보낸 code_challenge 의 원본으로, BASE64URL(SHA256(이 값)) 이 일치해야 한다.",
            "example": "dBjftJeZ4CVP-mB92K27uhbUJU1p1r_wW1gFWFOEjXk"
          }
        },
        "required": [
          "grant_type"
        ]
      },
      "TokenResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "access token(JWT)"
          },
          "tokenType": {
            "type": "string",
            "description": "토큰 타입",
            "example": "Bearer"
          },
          "expiresIn": {
            "type": "number",
            "description": "access token 만료(초)",
            "example": 3600
          },
          "refreshToken": {
            "type": "string",
            "description": "refresh token(불투명, rt_...). **1회용이다** — 이 값으로 갱신하면 새 refresh token 이 발급되고 이 값은 그 즉시 무효가 된다. 응답을 받으면 반드시 새 값으로 교체해 보관할 것."
          },
          "refreshExpiresAt": {
            "type": "string",
            "description": "refresh token 만료 시각(ISO8601)",
            "example": "2026-09-20T12:00:00.000Z"
          }
        },
        "required": [
          "accessToken",
          "tokenType",
          "expiresIn",
          "refreshToken",
          "refreshExpiresAt"
        ]
      },
      "AuthorizeRequestDto": {
        "type": "object",
        "properties": {
          "redirectUri": {
            "type": "string",
            "description": "OAuth2 redirect_uri — 코드를 실어 돌려보낼 클라이언트 URL. clientId 를 주면 그 클라이언트에 등록된 리디렉션 URI 와 정확히 일치해야 하고, 없으면 1st-party 허용목록 오리진이어야 한다.",
            "example": "https://medifinder.kr/auth/callback"
          },
          "clientId": {
            "type": "string",
            "description": "외부 앱의 공개 클라이언트 ID. 생략하면 1st-party(인증웹 자신)로 간주한다. 이 값은 발급되는 코드에 기록되어, 토큰 교환 때 요청 Origin 대조의 기준이 된다.",
            "example": "cl_fixed_medifinder"
          },
          "codeChallenge": {
            "type": "string",
            "description": "PKCE code_challenge = BASE64URL(SHA256(code_verifier)), 43자. S256 만 받는다. 발급되는 코드에 기록되어 교환 때 code_verifier 와 대조된다.",
            "example": "E9Melhoa2OwvFrEMTJguCHaoeK1t8URWbuGJSstw-cM"
          }
        },
        "required": [
          "redirectUri",
          "codeChallenge"
        ]
      },
      "AuthorizeResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "1회용 인가코드(ac_...)"
          }
        },
        "required": [
          "code"
        ]
      },
      "MeResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "회원번호"
          },
          "email": {
            "type": "string",
            "description": "이메일"
          },
          "emailVerified": {
            "type": "boolean",
            "description": "이메일 인증 여부"
          },
          "name": {
            "type": "object",
            "description": "표시 이름"
          },
          "role": {
            "type": "string",
            "description": "권한",
            "example": "USER"
          },
          "tier": {
            "type": "string",
            "description": "회원 등급. 앱 생성 한도를 정한다.",
            "enum": [
              "BASIC",
              "PRO",
              "UNLIMITED"
            ],
            "example": "BASIC"
          },
          "joinType": {
            "type": "string",
            "description": "최초 가입 수단",
            "example": "EMAIL"
          },
          "language": {
            "type": "string",
            "description": "표시·메일 언어. 고른 적이 없으면 null 이고, 그때는 요청의 Accept-Language 를 따른다.",
            "enum": [
              "ko",
              "en",
              "ja",
              "zh"
            ],
            "example": "ko"
          },
          "timeZone": {
            "type": "string",
            "description": "IANA 타임존 ID. 고른 적이 없으면 null 이다.",
            "example": "Asia/Seoul"
          },
          "createdAt": {
            "type": "string",
            "description": "가입 일시",
            "example": "2026-08-06T02:10:00.000Z"
          },
          "hasPassword": {
            "type": "boolean",
            "description": "비밀번호가 설정돼 있는가. 소셜로만 가입한 계정은 false 라 화면이 \"변경\" 대신 \"설정\" 을 띄운다."
          },
          "linkedProviders": {
            "description": "연동된 소셜 제공자 목록",
            "example": [
              "GOOGLE",
              "KAKAO"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "email",
          "emailVerified",
          "role",
          "tier",
          "joinType",
          "createdAt",
          "hasPassword",
          "linkedProviders"
        ]
      },
...
    }
  }
}